ndi-bridge/requirements.txt
# WebRTC
aiortc>=1.5.0
aiohttp>=3.8.0

# Video Processing
opencv-python>=4.8.0
numpy>=1.24.0
av>=10.0.0

# API Server
fastapi>=0.100.0
uvicorn>=0.20.0

# Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# WebSocket
websockets>=11.0

# NDI
ndi-python>=5.1.0

# Additional dependencies
asyncio-mqtt>=0.11.0

